refactor: 엔티티 캐릭터 채팅 구조 변경
